Use epoch and round to create the new block nonce.

+++ crates/consensus/primary/src/consensus/leader_schedule.rs
@@ -119,7 +119,7 @@ impl LeaderSwapTable {
     pub fn swap(&self, leader: &AuthorityIdentifier, leader_round: Round) -> Option<Authority> {
         if self.bad_nodes.contains_key(leader) {
             let mut seed_bytes = [0u8; 32];
-            seed_bytes[32 - 8..].copy_from_slice(&leader_round.to_le_bytes());
+            seed_bytes[32 - 8..].copy_from_slice(&(leader_round as u64).to_le_bytes());
             let mut rng = StdRng::from_seed(seed_bytes);
 
             let good_node = self
@@ -229,7 +229,7 @@ impl LeaderSchedule {
                 // we can always divide by 2 to get a monotonically incremented sequence,
                 // 2/2 = 1, 4/2 = 2, 6/2 = 3, 8/2 = 4  etc, and then do minus 1 so we can always
                 // start with base zero 0.
-                let next_leader = (round/2 + self.committee.size() as u64 - 1) as usize % self.committee.size();
+                let next_leader = (round as u64 / 2 + self.committee.size() as u64 - 1) as usize % self.committee.size();
 
                 let leader: Authority = self.committee.authorities().nth(next_leader).expect("authority out of bounds!").clone();
                 let table = self.leader_swap_table.read();
@@ -237,7 +237,7 @@ impl LeaderSchedule {
                 table.swap(&leader.id(), round).unwrap_or(leader)
             } else {
                 // Elect the leader in a stake-weighted choice seeded by the round
-                let leader = self.committee.leader(round);
+                let leader = self.committee.leader(round as u64);
 
                 let table = self.leader_swap_table.read();
                 table.swap(&leader.id(), round).unwrap_or(leader)
